Move Sound settings to `menus_settings_sound.cpp`
